Base and path generation methods implemented
The contents of this commit are a few methods used to define a (n x m) numerical array composed of different values, each one representing either a 'wall', 'path', 'initial node' or 'goal node'. This is done in order to provide with a completely random array generation that can be interpreted as a 'maze' for visual aid in the design of search algorithms.