Path gen optimization #1
The algorithm used to randomly generate paths in the maze ('1' values in the numerical array) has been optimized in terms of design.

The inner walls are now more squared, meaning that they are less spread out along the avaliable field of the array. However, there are two problems: the minor one, some paths have a multiple-tile width, which can cause inconsistency during the search algorithms execution. On the other hand, there is a major problem regarding generation speed. The time elapsed to properly generate the path takes more than 100 times longer than the previous algorithm.

That last issue might be related with duplicated value insertions in the arrays used for the generation method.