Path gen optimization #2
Minor improvement in terms of generation speed (removed the cloned array).

There is still, however, a considerable time delay due to the amount of iterations that the algorithm performs. These repeated iterations are due to the accumulation of coordinate values in the 'Frontier' array, which causes the repetitive execution of both 'nextnodes' and 'surroundings' methods, with their corresponding internal iterations.

The grand total of iterations performed by the whole process ascends to '3,999,567'.