Overall enhancements
This commit implements a revision and correction of the method's docstrings and comments along the code.

Some 'if-else' statements were condensed as ternary operators for general code coherence enhancement.

Some variables were renamed for the same reason as the previous case.

Moved 'display' method to the bottom of the file in order to preserve the operational structure of the class' objects call.