Solve Method Update #1
First solving method implemented (DFS), proven to be considerably inefficient.

Auxiliar debugging method implemented, 'log'. This will be removed when the algorithms design is done and checked.

Note: the '__init__()' method for the Solver class seems to be redundant and there might be a better way to produce the same output in a tidier way. However, this is not yet a preference.