Solve Method Update #2
Added second pathfinding method. This one selects a node from the frontier depending on its distance to the goal node. Picks the closest one and evaluates it.

Fixed some minor performance issues regarding comparison checks such as 'if node not in [explored, self.Initial, self.Goal]', which did not check if the node was in the 'explored' array, only 'if node == explored', which returned False every time, since 'type(node)' = 'tuple' and 'type(explored)' = 'list'.