Major changes
Redefined logging system: switched from a visual logging system to a file logging one.

Added an ASCII display method for the maze, as well as an image-generation method that provides with better visualization.

Added a feature that enables the nodes to be colored procedurally as a gradient, automatically adjusted given start and end nodes, as well as the amount of total explored ones.